Move mock interfaces into `resource_test.go`.
